Move portfolio api methods into module's folder
